pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "falconlogger"
version = "0.0.2"
authors = [
    { name="Thomas J. Daley", email="tom@powerdaley.com" },
]
description = "Consistent use of Python logger package"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
]
classifiers = [
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Natural Language :: English",
]

[project.urls]
"Homepage" = "https://github.com/tjdaley/falconlogger"
"Bug Tracker" = "https://github.com/tjdaley/falconlogger/issues"
"Developer's Blog" = "https://www.thomasjdaley.com"
